+++ app/src/main/java/com/example/nework/model/EventModel.kt
@@ -4,8 +4,8 @@ import com.example.nework.dto.Coords
 
 data class EventModel(
     val coords: Coords? = null,
-    val speakerIds: List<Int> = emptyList(),
-    val participantsIds: List<Int> = emptyList()
+    val speakerIds: List<Long> = emptyList(),
+    val participantsIds: List<Long> = emptyList()
 )
