+++ app/src/main/java/com/example/nework/model/PostModel.kt
@@ -4,5 +4,5 @@ import com.example.nework.dto.Coords
 
 data class PostModel(
     val coords: Coords? = null,
-    val mentionIds: List<Int> = emptyList()
+    val mentionIds: List<Long> = emptyList()
 )
