+++ app/src/main/java/com/example/nework/ui/GetUserListDialogFragment.kt
@@ -9,15 +9,12 @@ import androidx.core.os.bundleOf
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.asLiveData
-import androidx.navigation.fragment.findNavController
 import com.example.nework.R
 import com.example.nework.adapter.OnIterationUserListener
 import com.example.nework.adapter.UserAdapter
-import com.example.nework.api.AppApi
 import com.example.nework.auth.AppAuth
 import com.example.nework.databinding.DialogListUserBinding
 import com.example.nework.dto.User
-import com.example.nework.ui.UserFragment.Companion.USER_ID
 import com.example.nework.vm.UsersViewModel
 import com.example.nework.vm.UsersViewModelFactory
 import dagger.hilt.android.AndroidEntryPoint
@@ -28,7 +25,7 @@ import javax.inject.Inject
 @Suppress("UNCHECKED_CAST")
 @AndroidEntryPoint
 class GetUserListDialogFragment(
-    private val listIds: List<Int>,
+    private val listIds: List<Long>,
     private val titleId: Int
 ) : DialogFragment() {
     private var _listView: View? = null
